+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemTest.java
@@ -476,11 +476,6 @@ public class GoogleCloudStorageFileSystemTest
    * TODO(user): add support of generations in InMemoryGoogleCloudStorage so
    * we can run the following tests in this class.
    */
-  @Test
   @Override
-  public void testReadGenerationBestEffort() throws IOException {}
-
-  @Test
-  @Override
-  public void testReadGenerationStrict() throws IOException {}
+  public void read_failure_ifObjectWasModifiedDuringRead() {}
 }