+++ gcs/CHANGES.md
@@ -13,6 +13,9 @@
         fs.gs.storage.root.url
         fs.gs.token.server.url
 
+1.  Always set `generation` parameter for read requests and remove
+    `fs.gs.generation.read.consistency` property.
+
 ### 2.0.1 - 2020-02-13
 
 1.  Cooperative Locking FSCK tool: fix recovery of operations that failed before